requirements.txt
# Core FastAPI and server dependencies - compatible versions
fastapi==0.115.9
uvicorn[standard]==0.34.0
gunicorn==21.2.0
python-dotenv==1.0.1

# LangChain and LLM dependencies
langchain-core>=0.3.49
# Commented out Mistral AI dependencies
# langchain-mistralai==0.2.11
# mistralai==0.4.2
# Google Gemini API dependencies - use compatible versions
google-generativeai>=0.8.0,<0.9.0
langchain-google-genai>=1.0.0,<2.1.0

# Database dependencies
reportlab==4.0.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.38
sqlalchemy-utils==0.41.1

# Web framework dependencies
jinja2==3.1.2
python-multipart==0.0.18

# Web scraping dependencies - updated versions
requests==2.32.3
beautifulsoup4==4.12.2
lxml>=5.3.0
feedparser==6.0.10
praw==7.7.1
aiohttp==3.11.11
httpx>=0.27.0
selenium==4.15.2
fake-useragent==1.5.1

# Authentication dependencies - compatible versions
python-jose==3.4.0
passlib==1.7.4
bcrypt==4.3.0
email-validator==2.1.0
